src/test/kotlin/canon/parser/xml/strategy/SelectionStrategyTest.kt
package canon.parser.xml.strategy

import canon.extension.toNode
import canon.parser.xml.CanonXmlParser
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Assertions.assertNotNull
import org.junit.jupiter.api.Test

class SelectionStrategyTest {

    @Test
    fun testParse() {
        val xml = "<selection id='testId' class='testClass' name='testName'>" +
                "</selection>"
        val parsed = SelectionStrategy().parse(xml.toNode(), HashMap<String, Any?>(), CanonXmlParser()::toRenderables)

        assertNotNull(parsed)
        assertEquals("testId", parsed.id)
        assertEquals("testClass", parsed.`class`)
        assertEquals("testName", parsed.name)
        assertEquals(0, parsed.countdownInSec)
    }

}
